requirements.txt
fastapi==0.115.5
uvicorn[standard]==0.32.1
langchain>=0.3.10
langchain-community>=0.3.10
langchain-ollama>=0.2.2
langchain-google-genai>=2.0.8
langchain-core>=0.3.0
pydantic==2.10.3
python-dotenv==1.0.1
httpx==0.28.1

arxiv==2.1.3
feedparser==6.0.11
beautifulsoup4==4.12.3
requests==2.32.3

PyPDF2==3.0.1
pymupdf==1.24.0
pdfplumber==0.11.0

langchain-huggingface==0.1.2
sentence-transformers==3.3.1

pymongo==4.10.1
motor==3.6.0

gradio==5.9.1

nltk==3.9.1
scikit-learn==1.5.2
numpy==1.26.4
pandas==2.2.3

# VectorDB
faiss-cpu==1.9.0  # FAISS 벡터 데이터베이스
chromadb==0.5.18  # ChromaDB 벡터 데이터베이스 (대안)

# 하이브리드 Agentic Chunking
llama-cpp-python==0.2.90  # 로컬 LLM (CPU/GPU)
google-generativeai==0.8.3  # Gemini API

# MCP (Model Context Protocol)
mcp>=0.1.0

